adding authorization checks for delete/edit posts

+++ rest-backend/controllers/feed.js
@@ -132,6 +132,11 @@ exports.updatePost = (req, res, next) => {
             error.statusCode = 404;
             throw error;
         }
+        if(post.creator.toString() !== req.userId){
+            const error = new Error('Not authorized');
+            error.statusCode = 403;
+            throw error;
+        }
         //if image changes, delete old image
         if(imageUrl !== post.imageUrl) {
             clearImage(post.imageUrl);
@@ -165,7 +170,12 @@ exports.deletePost = (req, res, next) => {
             const error = new Error('Could not find post to update');
             error.statusCode = 404;
             throw error;
-        }        
+        }  
+        if(post.creator.toString() !== req.userId){
+            const error = new Error('Not authorized');
+            error.statusCode = 403;
+            throw error;
+        }      
         clearImage(post.imageUrl);
         return Post.findByIdAndRemove(postId);
     })
